ci(e2e): split update-dependents-cascade and mocha-tester for node balancing
