+++ scripts/e2e-test-timings.json
@@ -147,7 +147,9 @@
   "e2e/harmony/lanes/skip-publish-unpublished-dep.e2e.ts": 103,
   "e2e/harmony/lanes/switch-lanes.e2e.ts": 453,
   "e2e/harmony/lanes/unrelated-and-removed.e2e.ts": 37,
-  "e2e/harmony/lanes/update-dependents-cascade.e2e.ts": 862,
+  "e2e/harmony/lanes/update-dependents-cascade-import.e2e.ts": 340,
+  "e2e/harmony/lanes/update-dependents-cascade-reset.e2e.ts": 272,
+  "e2e/harmony/lanes/update-dependents-cascade.e2e.ts": 250,
   "e2e/harmony/link.e2e.ts": 62,
   "e2e/harmony/load-bit.e2e.ts": 15,
   "e2e/harmony/load-extensions.e2e.ts": 312,
@@ -156,7 +158,8 @@
   "e2e/harmony/log.e2e.ts": 15,
   "e2e/harmony/merge-config.e2e.ts": 644,
   "e2e/harmony/missing-package-strips-trailing-slash.e2e.ts": 70,
-  "e2e/harmony/mocha-tester.e2e.ts": 474,
+  "e2e/harmony/mocha-tester-spec-files.e2e.ts": 214,
+  "e2e/harmony/mocha-tester.e2e.ts": 260,
   "e2e/harmony/multiple-compilers.e2e.ts": 300,
   "e2e/harmony/multiple-envs.e2e.ts": 80,
   "e2e/harmony/multiple-testers.e2e.ts": 79,
